feat(Machine, MachineClip): 투입 금액 및 자판기 보유 금액 관리
Machine
- 손님이 투입한 금액을 관리한다.

MachineClip
- 보유하고 있는 동전의 갯수를 관리한다.
